requirements.txt
pandas
faker>=30.0.0,<=35.0.0
python-dotenv
faker
jsonschema
kafka-python
tenacity
cassandra-driver
# apache-flink==2.0.0
vnstock
bokeh
fastapi==0.103.0
Twisted==24.7.0
annotated-types==0.5.0
anyio==3.7.1
pyvi==0.1.1
asyncpg==0.28.0
click==8.1.7
colorama==0.4.6
joblib==1.3.2
py4j==0.10.9.7
pydantic==2.3.0
pydantic_core==2.6.3
sniffio==1.3.0
tqdm==4.66.1
update-checker==0.18.0
urllib3==2.0.4
websocket-client==1.6.2
uvicorn[standard]==0.23.2
pytz == 2024.2